[web] Move device selection to its own section
And take the opportunity to avoid triggering the onChange callback when
the user actually does not change the selected device but clicks the
"Accept" button of the selection dialog.
