+++ web/src/components/storage/ProposalSettingsSection.test.jsx
@@ -86,116 +86,6 @@ beforeEach(() => {
   props = {};
 });
 
-describe("Installation device field", () => {
-  describe("if it is loading", () => {
-    beforeEach(() => {
-      props.isLoading = true;
-    });
-
-    describe("and there is no selected device yet", () => {
-      beforeEach(() => {
-        props.settings = { bootDevice: "" };
-      });
-
-      it("renders a message indicating that the device is not selected", () => {
-        plainRender(<ProposalSettingsSection {...props} />);
-
-        screen.getByText(/Installation device/);
-        screen.getByText(/No device selected/);
-      });
-    });
-
-    // FIXME: skipping this test example by now.
-    // It will be addressed when reworking the Device section
-    // as part of https://github.com/openSUSE/agama/pull/1031
-    describe.skip("and there is a selected device", () => {
-      beforeEach(() => {
-        props.settings = { bootDevice: "/dev/vda" };
-      });
-
-      it("renders the selected device", () => {
-        plainRender(<ProposalSettingsSection {...props} />);
-
-        screen.getByText(/Installation device/);
-        screen.getByText("/dev/vda");
-      });
-    });
-  });
-
-  describe("if there is no selected device yet", () => {
-    beforeEach(() => {
-      props.settings = { bootDevice: "" };
-    });
-
-    it("renders a message indicating that the device is not selected", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      screen.getByText(/Installation device/);
-      screen.getByText(/No device selected/);
-    });
-  });
-
-  // FIXME: skipping this test example by now.
-  // It will be addressed when reworking the Device section
-  // as part of https://github.com/openSUSE/agama/pull/1031
-  describe.skip("if there is a selected device", () => {
-    beforeEach(() => {
-      props.settings = { bootDevice: "/dev/vda" };
-    });
-
-    it("renders the selected device", () => {
-      plainRender(<ProposalSettingsSection {...props} />);
-
-      screen.getByText(/Installation device/);
-      screen.getByText("/dev/vda");
-    });
-  });
-
-  it("allows selecting a device when clicking on the device name", async () => {
-    props = {
-      availableDevices: [vda],
-      settings: { bootDevice: "/dev/vda" },
-      onChange: jest.fn()
-    };
-
-    const { user } = plainRender(<ProposalSettingsSection {...props} />);
-
-    const button = screen.getByRole("button", { name: "/dev/vda, 1 KiB" });
-    await user.click(button);
-
-    const popup = await screen.findByRole("dialog");
-    within(popup).getByText("Installation device");
-
-    const accept = within(popup).getByRole("button", { name: "Accept" });
-    await user.click(accept);
-
-    expect(screen.queryByRole("dialog")).not.toBeInTheDocument();
-    expect(props.onChange).toHaveBeenCalled();
-  });
-
-  it("allows canceling the selection of the device", async () => {
-    props = {
-      availableDevices: [vda],
-      settings: { bootDevice: "/dev/vda" },
-      onChange: jest.fn()
-    };
-
-    const { user } = plainRender(<ProposalSettingsSection {...props} />);
-
-    const button = screen.getByRole("button", { name: "/dev/vda, 1 KiB" });
-    await user.click(button);
-
-    const popup = await screen.findByRole("dialog");
-    within(popup).getByText("Installation device");
-
-    const cancel = within(popup).getByRole("button", { name: "Cancel" });
-    await user.click(cancel);
-
-    expect(screen.queryByRole("dialog")).not.toBeInTheDocument();
-    expect(props.onChange).not.toHaveBeenCalled();
-  });
-});
-
 describe("LVM field", () => {
   describe("if LVM setting is not set yet", () => {
     beforeEach(() => {
