+++ web/src/components/storage/index.js
@@ -23,6 +23,7 @@ export { default as ProposalPage } from "./ProposalPage";
 export { default as ProposalPageMenu } from "./ProposalPageMenu";
 export { default as ProposalSettingsSection } from "./ProposalSettingsSection";
 export { default as ProposalSpacePolicySection } from "./ProposalSpacePolicySection";
+export { default as ProposalDeviceSection } from "./ProposalDeviceSection";
 export { default as ProposalActionsSection } from "./ProposalActionsSection";
 export { default as ProposalVolumes } from "./ProposalVolumes";
 export { default as DASDPage } from "./DASDPage";
